Neutron TOS
This is the terms of service of the Neutron Bot.
These Terms of Service (“Terms,” “ToS”) represent our rules, policy, and guidelines for using our discord bot known as Neutron.

ntroduction These Terms are a binding legal agreement between you the User, “User,” and us. By visiting or using our discord bot now known as our “Services” you agree you are 13 years or older and that you read, agree, and accept being bound by these Terms.

We reserve the rights to update these Terms at any time, with or without notice, for any reason including but not limited to changes in law and features being added to the website. Disputes over changes should be handled through an abuse and dispute report, as reads below. 

right to disable, change, or rebuild the Service temporarily or permanently at any time with or without notice. Access to these Services is a privilege and can be revoked at any time for any reason stated in our conduct policy. Bots and automated programs may access the Service, if in compliance with our developer policy and conduct policy.

These Services uphold a basic conduct policy as follows. You, the user, agree to:

Not fake information submitting to our services Not display toxic behavior to the Staff or other users Not attempt to, or succeed in, taking other users private information Not participate in any illegal activity on Discord or our services Not ignore or stand against non-invasive staff directives We respect that behavior outside our Services, unless potentially hazardous to our Services, shall not impact access or abilities on the Service.

Breaking these Terms could lead to being removed, temporarily or permanently, at any time.

Communications By agreeing to these Terms, the user agrees to open electronic communication between the Service and them. We have the right to collect your email and Discord information to contact you at any time for matters related to the Service. All non-priority communications, including but limited to marketing notices, will not be sent without prior consent.

The Service agrees to be open to communications from users to Moderators, “Staff,” or our Privacy & Policy Team. All disputes can be handled as read below in the Terms.

As a user or developer you agree to;

Follow our conduct policy, without exception Allow us to collect data you upload, or that is related to your Discord account Allow us to communicate with you as needed Moderate the content you upload or transmit through our Services Provide a notice to Privacy & Policy Team prior to legal action Not automate or reverse-engineer our website in a way that doesn’t comply with our above Terms Not use our website or Services until you are 13 years of age or older Warranty and Liability The Services provided are provided “as is” with no warranty or liability for any outcomes of using this Service and/or any content you may or may not interact with on our Service. Our Service is used at your own risk, and you agree we are not liable for past, current, or future events that are associated with our Service or involve content on our Service. You proceed on our platform with no protections except as provided in these Terms. You are agreeing to indemnify our Service against all possible situations that might arise, except as stated under our abuse policy.